requirements.txt
ultralytics==8.0.0
torch>=1.7.0
torchvision>=0.8.1
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
seaborn>=0.11.0
pandas>=1.1.4
wandb>=0.12.0
tqdm>=4.41.0
